==> twitter_sentiment_classifier/notebooks/3. Train SentimentModel.ipynb <==
%load_ext autoreload
%autoreload 2

%load_ext autotime

%cd ..

# Necessary to be in root to get access to store
%cd ..

# Train SentimentModel

This notebook will train the last `N` layers of the 12-layered RobBERT Transformer model.

import json
import os
import pandas as pd

from twitter_sentiment_classifier.store.loader import fetch_all_tweet_data
from tqdm import tqdm
from collections import Counter

## 1. Load in data

Load in the different datasets (`train`, `test`, and `validation`).

# Fetch all tweet data from S3 (skips automatically if this is already the case)
fetch_all_tweet_data()

# Training
with open(os.path.expanduser('twitter_sentiment_classifier/store/data/tweets_train.jsonl'), 'r') as f:
    annotations_train = [json.loads(line) for line in f.readlines()]
print(f"Loaded in {len(annotations_train)} training annotations")

# Testing
with open(os.path.expanduser('twitter_sentiment_classifier/store/data/tweets_test.jsonl'), 'r') as f:
    annotations_test = [json.loads(line) for line in f.readlines()]
print(f"Loaded in {len(annotations_test)} testing annotations")

# Validation
with open(os.path.expanduser('twitter_sentiment_classifier/store/data/tweets_val.jsonl'), 'r') as f:
    annotations_val = [json.loads(line) for line in f.readlines()]
print(f"Loaded in {len(annotations_val)} validation annotations")

## 2. Load model

Load in the model used to train. We choose to train the RobBERT model from scratch, to not introduce any biases in a previously trained model.

from twitter_sentiment_classifier.sentiment_model import SentimentModel

# Load in the model
model = SentimentModel(
    cuda=True,
)
model

# Start from a clean slate by resetting the model
# model.reset(initialise=False)  # Since destructive effect, comment (just to be sure)

## 3. Training

Train part of the model for a limited amount of epochs.

# Number of training epochs
N_EPOCHS = 2

# Number of last layers prone to change during training
N_LAYERS = 1

model.train(
    n_epochs=N_EPOCHS,
    train_samples=annotations_train,
    val_samples=annotations_val,
    test_samples=annotations_test,
    n_last_layers=N_LAYERS,
)

## 4. Evaluation

Evaluate this intermediate version of the model. It should already have a decent performance.

model.eval(annotations_test)
